fix(routine-notification): adding code fixes and detaching routine component from check in notification component
