+++ src/components/Notifications/CheckInNotifications/messages.ts
@@ -1,13 +1,8 @@
 import { defineMessages } from 'react-intl'
 
-type ExploreTeamPageMessage = 'routineTitle' | 'keyResultsTitle'
+type ExploreTeamPageMessage = 'keyResultsTitle'
 
 export default defineMessages<ExploreTeamPageMessage>({
-  routineTitle: {
-    defaultMessage: 'Rotinas',
-    id: '+avOYN',
-    description: 'The title of the first section of My Week notification tab.',
-  },
   keyResultsTitle: {
     defaultMessage: 'Resultados-chave',
     id: 'n8OKn4',
