+++ src/components/Notifications/RoutineNotification/messages.ts
@@ -1,0 +1,11 @@
+import { defineMessages } from 'react-intl'
+
+type ExploreTeamPageMessage = 'routineTitle'
+
+export default defineMessages<ExploreTeamPageMessage>({
+  routineTitle: {
+    defaultMessage: 'Rotinas',
+    id: '+avOYN',
+    description: 'The title of the first section of My Week notification tab.',
+  },
+})
